Change mboTensorOpMatVec interface.
This function should take raw pointers to the x and y arrays to decouple from
MboVec.  In addition, it needs to take a range of rows [rmin, rmax) to make it
more consistent with the other MboTensorOp methods and to allow chunking of rows
for better cache utilization and threading.

Examples compile and run successfully.